requirements.txt
h5py==2.9.0
imageio==2.5.0
ipython==7.7.0
matplotlib==3.1.1
numpy==1.17.0
opencv-python==4.1.0.25
Pillow==6.1.0
scikit-image==0.15.0
scikit-learn==0.21.3
scikit-video==1.1.11
scipy==1.3.1
sklearn==0.0
torch==0.4.1
torchfile==0.1.0
torchvision==0.2.1